- Corrected a conversion of types for the Engine. Unsafe conversions marked as AssertFatal( ..., "Huge data." ).
- Fixed a slip in the codeBlock case OP_ITER_BEGIN.

- Excluded OggTheoraDecoder::_transcode420toRGBA_SSE2(). TODO is added.

- Fixed Con::printf() format for pointers.

+++ Engine/source/platformWin32/winConsole.cpp
@@ -147,7 +147,7 @@ void WinConsole::printf(const char *s, ...)
    // Axe the color characters.
    Con::stripColorChars(buffer);
    // Print it.
-   WriteFile(stdOut, buffer, strlen(buffer), &bytes, NULL);
+   WriteFile(stdOut, buffer, (DWORD)strlen(buffer), &bytes, NULL);
    FlushFileBuffers( stdOut );
 }
